migrate cache test plan to use in-mem mongo, redis
